Added 5540 barometer implementation
Not sure if it works, but committing so that the code is out there.

The main issue would be the 32.768 KHz timer that the barometer needs.
It is implemented but will need to be verified with a scope. Might need
to switch to register A instead of C. Worst case, look at using RTC,
better timer library, or toggle the pin in the ISR. Barometer is ok
with slightly wonky frequency.